feat: move purchase order models to purchasing app
This commit moves the PurchaseOrder, PurchaseOrderLine, and
PurchaseOrderSupplierQuote models from the workflow app to a new
purchasing app. This change is made to improve the organization
and maintainability of the codebase by grouping related models
into a dedicated app.

The migration process involves:
- Creating a new purchasing app with the necessary models and
  migrations.
- Moving the existing database tables to the new app using
  RenameTable operations in the migration files.
- Creating proxy models in the workflow app to maintain backward
  compatibility for existing code that imports the models from
  the workflow app.
- Updating the INSTALLED_APPS setting to include the new
  purchasing app.
- Removing the original models from the workflow app.
