+++ apps/purchasing/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from . import views
+
+app_name = 'purchasing'
+
+urlpatterns = [
+    # Purchase order URLs will be moved here
+]
